ConfigArithParams协议更新 两个模块合一
