+++ SIFP/SIFP.sln
@@ -41,10 +41,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "License", "SIFP\License\Lic
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CreateLicense", "SIFP\CreateLicense\CreateLicense.csproj", "{89609661-9AAB-4DC4-8F9F-4637BA853F61}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConfigCorrection", "SIFP\Modules\ConfigCorrection\ConfigCorrection.csproj", "{ACFED5FE-272C-496C-80ED-25F17DE715AD}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConfigPostProc", "SIFP\Modules\ConfigPostProc\ConfigPostProc.csproj", "{6BE172EF-B778-4C29-A104-8C39814A9DBC}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VcselDriverDialog", "SIFP\Modules\VcselDriverDialog\VcselDriverDialog.csproj", "{E8289E99-E197-441A-8425-4541CF589BDF}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PasswordDialog", "SIFP\Modules\PasswordDialog\PasswordDialog.csproj", "{FCA9F23B-9679-4762-A757-D39EE75147FE}"
@@ -52,6 +48,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConifgAlg", "SIFP\Modules\ConifgAlg\ConifgAlg.csproj", "{628D4A73-617B-44C8-B65F-1D4623ACDCC9}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DummyServer", "SIFP\Tests\DummyServer\DummyServer.csproj", "{99F1AB79-2D61-4B16-9A07-FCF33F5F58B2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConfigArithParams", "SIFP\Modules\ConfigArithParams\ConfigArithParams.csproj", "{4631B66A-8220-446D-8E80-D071C14E82B7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -123,14 +121,6 @@ Global
 		{89609661-9AAB-4DC4-8F9F-4637BA853F61}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{89609661-9AAB-4DC4-8F9F-4637BA853F61}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{89609661-9AAB-4DC4-8F9F-4637BA853F61}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ACFED5FE-272C-496C-80ED-25F17DE715AD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ACFED5FE-272C-496C-80ED-25F17DE715AD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ACFED5FE-272C-496C-80ED-25F17DE715AD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ACFED5FE-272C-496C-80ED-25F17DE715AD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6BE172EF-B778-4C29-A104-8C39814A9DBC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6BE172EF-B778-4C29-A104-8C39814A9DBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6BE172EF-B778-4C29-A104-8C39814A9DBC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6BE172EF-B778-4C29-A104-8C39814A9DBC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E8289E99-E197-441A-8425-4541CF589BDF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E8289E99-E197-441A-8425-4541CF589BDF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E8289E99-E197-441A-8425-4541CF589BDF}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -147,6 +137,10 @@ Global
 		{99F1AB79-2D61-4B16-9A07-FCF33F5F58B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{99F1AB79-2D61-4B16-9A07-FCF33F5F58B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{99F1AB79-2D61-4B16-9A07-FCF33F5F58B2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4631B66A-8220-446D-8E80-D071C14E82B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4631B66A-8220-446D-8E80-D071C14E82B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4631B66A-8220-446D-8E80-D071C14E82B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4631B66A-8220-446D-8E80-D071C14E82B7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -164,12 +158,11 @@ Global
 		{72777311-814A-4725-80B6-0A567930C348} = {2CC5DFBF-A81F-459D-A3F6-CD8D16DC551A}
 		{1B143FA8-C252-46E3-9A48-B9CDFD04EBBF} = {2CC5DFBF-A81F-459D-A3F6-CD8D16DC551A}
 		{0CE4E406-88F1-47BA-BB5F-1F8B939892FB} = {2CC5DFBF-A81F-459D-A3F6-CD8D16DC551A}
-		{ACFED5FE-272C-496C-80ED-25F17DE715AD} = {2CC5DFBF-A81F-459D-A3F6-CD8D16DC551A}
-		{6BE172EF-B778-4C29-A104-8C39814A9DBC} = {2CC5DFBF-A81F-459D-A3F6-CD8D16DC551A}
 		{E8289E99-E197-441A-8425-4541CF589BDF} = {2CC5DFBF-A81F-459D-A3F6-CD8D16DC551A}
 		{FCA9F23B-9679-4762-A757-D39EE75147FE} = {2CC5DFBF-A81F-459D-A3F6-CD8D16DC551A}
 		{628D4A73-617B-44C8-B65F-1D4623ACDCC9} = {2CC5DFBF-A81F-459D-A3F6-CD8D16DC551A}
 		{99F1AB79-2D61-4B16-9A07-FCF33F5F58B2} = {81F0CFB8-65FF-432F-A24B-AA375F426C80}
+		{4631B66A-8220-446D-8E80-D071C14E82B7} = {2CC5DFBF-A81F-459D-A3F6-CD8D16DC551A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BE17A21F-6118-4EB9-BD8F-BA1574DA4920}
