+++ SIFP/SIFP/Modules/ConfigArithParams/ConfigArithParams.csproj
@@ -2,7 +2,7 @@
   <PropertyGroup>
     <TargetFramework>net5.0-windows</TargetFramework>
     <UseWPF>true</UseWPF>
-    <AssemblyName>ConfigCorrection</AssemblyName>
+    <AssemblyName>ConfigArithParams</AssemblyName>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="MaterialDesignThemes" Version="4.0.0" />
