--- SIFP/SIFP/Modules/ConfigCorrection/ConfigCorrectionModule.cs
@@ -1,20 +1,0 @@
-﻿using ConfigCorrection.Views;
-using Prism.Ioc;
-using Prism.Modularity;
-using Prism.Regions;
-
-namespace ConfigCorrection
-{
-    public class ConfigCorrectionModule : IModule
-    {
-        public void OnInitialized(IContainerProvider containerProvider)
-        {
-
-        }
-
-        public void RegisterTypes(IContainerRegistry containerRegistry)
-        {
-
-        }
-    }
-}
