--- SIFP/SIFP/Modules/ConfigPostProc/ConfigPostProc.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
-  <PropertyGroup>
-    <TargetFramework>net5.0-windows</TargetFramework>
-    <UseWPF>true</UseWPF>
-    <AssemblyName>ConfigPostProc</AssemblyName>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="MaterialDesignThemes" Version="4.0.0" />
-    <PackageReference Include="Prism.Wpf" Version="8.0.0.1909" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\Services\ICommunicationServices\Services.Interfaces.csproj" />
-    <ProjectReference Include="..\..\SIFP.Core\SIFP.Core.csproj" />
-  </ItemGroup>
-</Project>
