--- SIFP/SIFP/Modules/ConfigPostProc/Views/ConfigPostProcView.xaml
@@ -1,125 +1,0 @@
-﻿<UserControl
-    x:Class="ConfigPostProc.Views.ConfigPostProcView"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:core="clr-namespace:SIFP.Core;assembly=SIFP.Core"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:ConfigPostProc.Views"
-    xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:prism="http://prismlibrary.com/"
-    d:DesignHeight="300"
-    d:DesignWidth="300"
-    prism:ViewModelLocator.AutoWireViewModel="True"
-    mc:Ignorable="d">
-    <ScrollViewer VerticalScrollBarVisibility="Auto">
-        <ScrollViewer.Resources>
-            <Style BasedOn="{StaticResource MaterialDesignScrollBarMinimal}" TargetType="ScrollBar" />
-            <Style BasedOn="{StaticResource MaterialDesignCardGroupBox}" TargetType="GroupBox" />
-        </ScrollViewer.Resources>
-        <StackPanel>
-            <GroupBox
-                Margin="10"
-                Header="OutPut"
-                Visibility="{Binding IsExpert, Converter={StaticResource BooleanToVisibilityConverter}}">
-                <StackPanel Orientation="Vertical">
-                    <CheckBox
-                        Margin="0,5,0,5"
-                        Content="OutPointCloud"
-                        IsChecked="{Binding OutPointCloud}" />
-                    <CheckBox
-                        Margin="0,5,0,5"
-                        Content="OutConfidence"
-                        IsChecked="{Binding OutConfidence}" />
-                    <CheckBox
-                        Margin="0,5,0,5"
-                        Content="OutFlag"
-                        IsChecked="{Binding OutFlag}" />
-                    <ComboBox
-                        Margin="0,5,0,5"
-                        materialDesign:TextFieldAssist.PrefixText="PointCloudType : "
-                        ItemsSource="{x:Static core:EnumsTypes.PointCloudTypes}"
-                        SelectedItem="{Binding OutPointCloudType}" />
-                    <ComboBox
-                        Margin="0,5,0,5"
-                        materialDesign:TextFieldAssist.PrefixText="DepthValueType : "
-                        ItemsSource="{x:Static core:EnumsTypes.DepthValueTypes}"
-                        SelectedItem="{Binding OutDepthValueType}" />
-                    <ComboBox
-                        Margin="0,5,0,5"
-                        materialDesign:TextFieldAssist.PrefixText="DepthDataType : "
-                        ItemsSource="{x:Static core:EnumsTypes.DepthDataTypes}"
-                        SelectedItem="{Binding OutDepthDataType}" />
-                </StackPanel>
-            </GroupBox>
-
-            <GroupBox Margin="10" Header="Denoising">
-                <StackPanel Orientation="Vertical">
-                    <ComboBox
-                        Margin="0,5,0,5"
-                        materialDesign:TextFieldAssist.PrefixText="TDenoiseLevel : "
-                        ItemsSource="{x:Static core:EnumsTypes.DenoiseLevels}"
-                        SelectedItem="{Binding TDenoiseLevel}" />
-                    <ComboBox
-                        Margin="0,5,0,5"
-                        materialDesign:TextFieldAssist.PrefixText="SDenoiseLevel : "
-                        ItemsSource="{x:Static core:EnumsTypes.DenoiseLevels}"
-                        SelectedItem="{Binding SDenoiseLevel}" />
-                    <ComboBox
-                        Margin="0,5,0,5"
-                        materialDesign:TextFieldAssist.PrefixText="SDenoiseMethod : "
-                        ItemsSource="{x:Static core:EnumsTypes.SDenoiseMethods}"
-                        SelectedItem="{Binding SDenoiseMethod}" />
-                </StackPanel>
-            </GroupBox>
-
-            <GroupBox
-                Margin="10"
-                Header="Repair"
-                Visibility="{Binding IsExpert, Converter={StaticResource BooleanToVisibilityConverter}}">
-                <StackPanel Orientation="Vertical">
-                    <CheckBox
-                        Margin="0,5,0,5"
-                        Content="DeFlyPixel"
-                        IsChecked="{Binding DeFlyPixel}" />
-
-                    <CheckBox
-                        Margin="0,5,0,5"
-                        Content="DeHoles"
-                        IsChecked="{Binding DeHoles}" />
-                </StackPanel>
-            </GroupBox>
-
-            <GroupBox
-                Margin="10"
-                Header="AntiInterference"
-                Visibility="{Binding IsExpert, Converter={StaticResource BooleanToVisibilityConverter}}">
-                <StackPanel Orientation="Vertical">
-                    <CheckBox
-                        Margin="0,5,0,5"
-                        Content="AntiALI"
-                        IsChecked="{Binding AntiALI}" />
-                    <CheckBox
-                        Margin="0,5,0,5"
-                        Content="AntiMCI"
-                        IsChecked="{Binding AntiMCI}" />
-                </StackPanel>
-            </GroupBox>
-
-            <GroupBox Margin="10" Header="Confidence">
-                <StackPanel Orientation="Vertical">
-                    <TextBox
-                        Margin="0,5,0,5"
-                        materialDesign:TextFieldAssist.PrefixText="ValidDistMin : "
-                        materialDesign:TextFieldAssist.SuffixText="mm"
-                        Text="{Binding ValidDistMin, UpdateSourceTrigger=PropertyChanged, ValidatesOnExceptions=True}" />
-                    <TextBox
-                        Margin="0,5,0,5"
-                        materialDesign:TextFieldAssist.PrefixText="ValidDistMax : "
-                        materialDesign:TextFieldAssist.SuffixText="mm"
-                        Text="{Binding ValidDistMax, UpdateSourceTrigger=PropertyChanged, ValidatesOnExceptions=True}" />
-                </StackPanel>
-            </GroupBox>
-        </StackPanel>
-    </ScrollViewer>
-</UserControl>
