+++ SIFP/SIFP/SIFP/ConfigViewTypes.cs
@@ -1,6 +1,5 @@
-﻿using ConfigCamera.Views;
-using ConfigCorrection.Views;
-using ConfigPostProc.Views;
+﻿using ConfigArithParams.Views;
+using ConfigCamera.Views;
 using ConifgAlg.Views;
 using System;
 
@@ -9,8 +8,7 @@ namespace SIFP
     public static class ConfigViewTypes
     {
         public static readonly Type ConfigCameraView = typeof(ConfigCameraView);
-        public static readonly Type ConfigCorrectionView = typeof(ConfigCorrectionView);
-        public static readonly Type ConfigPostProcView = typeof(ConfigPostProcView);
+        public static readonly Type ConfigArithParamsView = typeof(ConfigArithParamsView);
         public static readonly Type ConfigAlgView = typeof(ConfigAlgView);
     }
 }
